pyproject.toml
[tool.poetry]
name = "konstructcore"
version = "0.0.1"
description = "Cross-platform data types and async building blocks to empower the high-level konstruct business logic. Require Python 3.9+."
authors = ["A Pal <aplpl@pm.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9,<=3.14"
yapf = "^0.43.0"
mypy = "^1.13.0"
aiohttp = "^3.11.7"
more-itertools = "^10.5.0"
tzlocal = "^5.2"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
coverage = "^7.6.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
